% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{break_down_cond}
\alias{break_down_cond}
\title{Break down a condition}
\usage{
break_down_cond(cond, env = parent.frame(), names = TRUE)
}
\arguments{
\item{cond}{expression}

\item{env}{environment}

\item{names}{boolean}
}
\description{
Used by \code{verbose_if()} and \code{verbose_while()} and exported for convenience
}
