% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_time_once.R
\name{system_time_once}
\alias{system_time_once}
\title{Print the system.time of a Function's Next Call}
\usage{
system_time_once(f, gcFirst = TRUE)
}
\arguments{
\item{f}{function}

\item{gcFirst}{Logical - should a garbage collection be performed
    immediately before the timing?  Default is \code{TRUE}.}
}
\value{
invisible()
}
\description{
This doesn't affect the function's output, which would not be the case
if we ran \code{surround_once(f, system.time, gcFirst)}
}
\examples{
system_time_once(sample)
x <- sample(1e6)
}
