% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{verbose_for}
\alias{verbose_for}
\alias{verbose_if}
\alias{verbose_while}
\title{verbose control flow}
\usage{
verbose_for(fun = identity)

verbose_if()

verbose_while()
}
\arguments{
\item{fun}{a function to apply on \code{for}'s iterator}
}
\description{
These functions make \code{for}, \code{while} or \code{if} more talkative for their next call.
\code{verbose_for()} will print the iterator, or the latter transformed by the \code{fun}
argument. \code{verbose_if()} and \code{verbose_while()} will break down the boolean
condition and display it as a one row \code{data.frame}.
}
\examples{
roses_are_red <- TRUE
violets_are_blue <- TRUE
love <- 1i
verbose_if()
if (is.logical(love) || roses_are_red && violets_are_blue)
  print("Sugar is sweet my love, but not as sweet as you.")

x <- 1
y <- 2
verbose_while()
while(x < 5 && y > 0) x <- x + 1

verbose_for(dim)
l <- list(iris, cars)
for (x in l) {
print(head(x,1))
}
}
